move `model` & `db` to internal
