+++ libra/src/internal/mod.rs
@@ -1,5 +1,7 @@
-pub mod index;
-pub mod head;
 pub mod branch;
-pub mod protocol;
 pub mod config;
+pub mod db;
+pub mod head;
+pub mod index;
+pub mod model;
+pub mod protocol;
